+++ cognee/tasks/ingestion/ingest_data.py
@@ -1,14 +1,17 @@
 import dlt
 import cognee.modules.ingestion as ingestion
-
+from typing import Any, Union, BinaryIO
+from llama_index.core import Document
 from cognee.shared.utils import send_telemetry
 from cognee.modules.users.models import User
-from cognee.infrastructure.databases.relational import get_relational_config, get_relational_engine
+from cognee.infrastructure.databases.relational import get_relational_engine
+from cognee.modules.ingestion import save_data_to_file
+from .transform_data import get_data_from_llama_index
 from cognee.modules.data.methods import create_dataset
 from cognee.modules.users.permissions.methods import give_permission_on_document
 from .get_dlt_destination import get_dlt_destination
 
-async def ingest_data(file_paths: list[str], dataset_name: str, user: User):
+async def ingest_data_with_metadata(data: Any, dataset_name: str, user: User):
     destination = get_dlt_destination()
 
     pipeline = dlt.pipeline(
@@ -16,9 +19,39 @@ async def ingest_data(file_paths: list[str], dataset_name: str, user: User):
         destination = destination,
     )
 
+    def save_data_to_storage(data_item: Union[BinaryIO, Document, str], dataset_name: str) -> str:
+        # Check if data is of type Document or any of it's subclasses
+        if isinstance(data_item, Document):
+            file_path = get_data_from_llama_index(data_item, dataset_name)
+
+        # data is a file object coming from upload.
+        elif hasattr(data_item, "file"):
+            file_path = save_data_to_file(data_item.file, dataset_name, filename=data_item.filename)
+
+        elif isinstance(data_item, str):
+            # data is a file path
+            if data_item.startswith("file://") or data_item.startswith("/"):
+                file_path = data_item.replace("file://", "")
+            # data is text
+            else:
+                file_path = save_data_to_file(data_item, dataset_name)
+        else:
+            raise ValueError(f"Data type not supported: {type(data_item)}")
+
+        return file_path
+
     @dlt.resource(standalone = True, merge_key = "id")
-    async def data_resources(file_paths: str, user: User):
-        for file_path in file_paths:
+    async def data_resources(data: Any, user: User):
+        if not isinstance(data, list):
+            # Convert data to a list as we work with lists further down.
+            data = [data]
+
+        # Process data
+        for data_item in data:
+
+            file_path = save_data_to_storage(data_item, dataset_name)
+
+            # Ingest data and add metadata
             with open(file_path.replace("file://", ""), mode = "rb") as file:
                 classified_data = ingestion.classify(file)
 
@@ -34,20 +67,20 @@ async def ingest_data(file_paths: list[str], dataset_name: str, user: User):
                 async with db_engine.get_async_session() as session:
                     dataset = await create_dataset(dataset_name, user.id, session)
 
-                    data = (await session.execute(
+                    data_point = (await session.execute(
                         select(Data).filter(Data.id == data_id)
                     )).scalar_one_or_none()
 
-                    if data is not None:
-                        data.name = file_metadata["name"]
-                        data.raw_data_location = file_metadata["file_path"]
-                        data.extension = file_metadata["extension"]
-                        data.mime_type = file_metadata["mime_type"]
+                    if data_point is not None:
+                        data_point.name = file_metadata["name"]
+                        data_point.raw_data_location = file_metadata["file_path"]
+                        data_point.extension = file_metadata["extension"]
+                        data_point.mime_type = file_metadata["mime_type"]
 
-                        await session.merge(data)
+                        await session.merge(data_point)
                         await session.commit()
                     else:
-                        data = Data(
+                        data_point = Data(
                             id = data_id,
                             name = file_metadata["name"],
                             raw_data_location = file_metadata["file_path"],
@@ -55,7 +88,7 @@ async def ingest_data(file_paths: list[str], dataset_name: str, user: User):
                             mime_type = file_metadata["mime_type"],
                         )
 
-                        dataset.data.append(data)
+                        dataset.data.append(data_point)
                         await session.commit()
 
                 yield {
@@ -72,7 +105,7 @@ async def ingest_data(file_paths: list[str], dataset_name: str, user: User):
 
     send_telemetry("cognee.add EXECUTION STARTED", user_id = user.id)
     run_info = pipeline.run(
-        data_resources(file_paths, user),
+        data_resources(data, user),
         table_name = "file_metadata",
         dataset_name = dataset_name,
         write_disposition = "merge",
